strip whitespace from page labels
Testing this in a reasonable way isn't easy since the internals of how
we get metadata from the epub are tied into the metadata extractor.
Longer term we should probably add a seam between whatever gets metadata
from the epub and what formats that into the output metadata, which will
make this easier to test.