construct path to items in toc ourselves
This avoids a bug in epub-parser where it can't match the items referenced
in the toc to the items in the manifest. There is an open PR with a
partial fix for that issue, but it still fails in certain cases where
there's path traversal in the hrefs in the contents due to the way
Addressable::URI handles normalization: see
https://gitlab.com/KitaitiMakoto/epub-parser/merge_requests/1 for more
details.

There is still an issue where if the manifest is at OEBPS/content.opf,
the contents is at OEBPS/toc.xhtml, and the contents has hrefs like
../OEBPS/chapter01.xhtml (this is still valid according to epubcheck)
the resulting hrefs get formatted like "OEBPS/../OEBPS/chapter01.xhtml",
but that's probably OK for now - none of the test epubs I'm working with
have that pattern.